+++ packages/plugin-packages/mashroom-security-provider-ldap/src/mashroom-bootstrap.ts
@@ -7,8 +7,8 @@ import type {MashroomSecurityProviderPluginBootstrapFunction} from '@mashroom/ma
 
 const bootstrap: MashroomSecurityProviderPluginBootstrapFunction = async (pluginName, pluginConfig, pluginContextHolder) => {
     const {
-        loginPage, userSearchFilter, groupSearchFilter, extraDataMapping, groupToRoleMapping: groupToRoleMappingPath,
-        serverUrl, ldapConnectTimeout, ldapTimeout,
+        loginPage, userSearchFilter, groupSearchFilter, extraDataMapping, secretsMapping,
+        groupToRoleMapping: groupToRoleMappingPath, serverUrl, ldapConnectTimeout, ldapTimeout,
         tlsOptions, baseDN, bindDN, bindCredentials, authenticationTimeoutSec,
     } = pluginConfig;
 
@@ -22,7 +22,7 @@ const bootstrap: MashroomSecurityProviderPluginBootstrapFunction = async (plugin
        ldapClient.shutdown();
     });
 
-    return new MashroomLdapSecurityProvider(loginPage, userSearchFilter, groupSearchFilter, extraDataMapping,
+    return new MashroomLdapSecurityProvider(loginPage, userSearchFilter, groupSearchFilter, extraDataMapping, secretsMapping,
         groupToRoleMappingPath, ldapClient, serverRootFolder, authenticationTimeoutSec, loggerFactory);
 };