+++ packages/plugin-packages/mashroom-security-provider-openid-connect/test/MashroomOpenIDConnectSecurityProvider.test.ts
@@ -1,6 +1,7 @@
 
 import MashroomOpenIDConnectSecurityProvider from '../src/security-provider/MashroomOpenIDConnectSecurityProvider';
 import {OICD_AUTH_DATA_SESSION_KEY, OICD_USER_SESSION_KEY} from '../src/constants';
+import {MashroomSecurityRoles} from '@mashroom/mashroom-security/type-definitions';
 
 describe('MashroomOpenIDConnectSecurityProvider', () => {
 
@@ -12,14 +13,37 @@ describe('MashroomOpenIDConnectSecurityProvider', () => {
             session: {
                 [OICD_AUTH_DATA_SESSION_KEY]: {
                     tokenSet: {
+                        access_token: 'xxxxxxx',
                         expires_at: Date.now() / 1000 + 2000,
                     }
                 },
-                [OICD_USER_SESSION_KEY]: {}
+                [OICD_USER_SESSION_KEY]: {
+                    username: 'test',
+                    displayName: 'Test User',
+                    email: 'test@test.com',
+                    pictureUrl: null,
+                    extraData: {
+                        firstName: 'John'
+                    },
+                    roles: [],
+                }
             }
         };
 
         expect(provider.getUser(req)).toBeTruthy();
+        expect(provider.getUser(req)).toEqual({
+            displayName: 'Test User',
+            email: 'test@test.com',
+            extraData: {
+                firstName: 'John'
+            },
+            pictureUrl: null,
+            roles: [],
+            secrets: {
+                accessToken: 'xxxxxxx'
+            },
+            username: 'test'
+        });
     });
 
     it('does not return the user if the token is expired', () => {